feat: introduce lazy-loading → memo page
